src/main/webapp/app/entities/tache/route/tache-routing-resolve.service.spec.ts
jest.mock('@angular/router');

import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, Router } from '@angular/router';
import { of } from 'rxjs';

import { ITache, Tache } from '../tache.model';
import { TacheService } from '../service/tache.service';

import { TacheRoutingResolveService } from './tache-routing-resolve.service';

describe('Tache routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: TacheRoutingResolveService;
  let service: TacheService;
  let resultTache: ITache | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      providers: [Router, ActivatedRouteSnapshot],
    });
    mockRouter = TestBed.inject(Router);
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRouteSnapshot);
    routingResolveService = TestBed.inject(TacheRoutingResolveService);
    service = TestBed.inject(TacheService);
    resultTache = undefined;
  });

  describe('resolve', () => {
    it('should return ITache returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 'ABC' };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultTache = result;
      });

      // THEN
      expect(service.find).toBeCalledWith('ABC');
      expect(resultTache).toEqual({ id: 'ABC' });
    });

    it('should return new ITache if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultTache = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultTache).toEqual(new Tache());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Tache })));
      mockActivatedRouteSnapshot.params = { id: 'ABC' };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultTache = result;
      });

      // THEN
      expect(service.find).toBeCalledWith('ABC');
      expect(resultTache).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
